==> 01-data-cleaning-election-result.ipynb <==
## Data Cleaning

#### Import modules

import os
import pandas as pd

#### Define paths and load data

pd.set_option('display.max_columns', None)

PATH = 'raw_data'

OUTPUT_PATH = 'outputs'

### Election results

#### Load data

results = pd.read_csv(os.path.join(PATH, 'HarvardData_countypres_2000-2020.csv'))
results = results[results['party'].isin(['DEMOCRAT', 'REPUBLICAN'])]
grpby_cols = ['year', 'state', 'state_po', 'county_name', 'county_fips', 'office',
              'candidate', 'party', 'totalvotes', 'version']
results = results.groupby(grpby_cols, as_index=False).sum()

results['pctvotes'] = results['candidatevotes'] / results['totalvotes']
results['office'] = 'PRESIDENT'  # for consistency

results.state.unique()

#### Pivot table

index_cols = ['year', 'state', 'state_po', 'county_name', 
              'county_fips', 'office']

pivoted_results = pd.pivot_table(results, index=index_cols, values=['pctvotes', 'candidatevotes'], columns=['party']).reset_index()
pivoted_results.columns.name = ''
pivoted_results.columns = [(x[0] + '_' + x[1]).lower() if x[1] != '' else x[0] for x in  pivoted_results.columns.values]
a = pivoted_results['candidatevotes_democrat'] > pivoted_results['candidatevotes_republican']
pivoted_results['winner'] = ['democrat' if x else 'republican' for x in a]
pivoted_results = pivoted_results[['year', 'state', 'county_name', 'county_fips', 'candidatevotes_democrat',
                                   'candidatevotes_republican', 'pctvotes_democrat', 'pctvotes_republican', 'winner']]

pivoted_results.head()

Export election results

for year in pivoted_results['year'].unique():
    df_temp = pivoted_results[pivoted_results['year'] == year]
    df_temp.to_csv(os.path.join(OUTPUT_PATH, f'results_{year}.csv'), index=False)

### Unemployment data

unemployment = pd.read_excel(os.path.join(PATH, 'ERS_Unemployment.xlsx'), header=4)

rates_cols = ['Unemployment_rate_2000',
              'Unemployment_rate_2004',
              'Unemployment_rate_2008',
              'Unemployment_rate_2012',
              'Unemployment_rate_2016',
              'Unemployment_rate_2020']

area_cols = ['State', 'Area_name', 'FIPS_Code']

unemployment_ey = unemployment[area_cols + rates_cols]

unemployment_ey.sample(2)

unemployment_ey_melted = pd.melt(unemployment_ey, 
                                 id_vars=area_cols, 
                                 value_vars=rates_cols, 
                                 value_name='unemployment',
                                 var_name='year')
unemployment_ey_melted['year'] = [x.split('_')[2] for x in unemployment_ey_melted['year']]
unemployment_ey_melted['year'] = unemployment_ey_melted['year'].astype(int)
unemployment_ey_melted.rename(columns={'FIPS_Code': 'county_fips'}, inplace=True)
unemployment_ey_melted = unemployment_ey_melted[['county_fips', 'year', 'unemployment']]

unemployment_ey_melted.sample()

Export unemployment data

for year in unemployment_ey_melted['year'].unique():
    
    df_temp = unemployment_ey_melted[unemployment_ey_melted['year'] == year]
    df_temp.to_csv(os.path.join(OUTPUT_PATH, f'unemployment_{year}.csv'), index=False)

#### Population

population = pd.read_excel(os.path.join(PATH, 'ERS_PopulationEstimates.xls'), header=2)

population.sample()

We obtained the rates for natural increase in population, international migration, domestic migration, and net migration for the years 2012, 2016, and we substituted the value of 2019 for 2020. 

years_population = ['2012', '2016', '2019']

rate_nat_inc_cols = ['R_NATURAL_INC_' + str(x) for x in years_population]
rate_int_mig_cols = ['R_INTERNATIONAL_MIG_' + str(x) for x in years_population]
rate_dom_mig_cols = ['R_DOMESTIC_MIG_' + str(x) for x in years_population]
rate_net_mig_cols = ['R_NET_MIG_' + str(x) for x in years_population]

population_dict = {'rate_natural_increase_population': rate_nat_inc_cols,
                   'rate_international_migration': rate_int_mig_cols,
                   'rate_domestic_migration': rate_dom_mig_cols,
                   'rate_net_migration': rate_net_mig_cols}

for key in population_dict.keys():
    
    cols = population_dict[key]
    temp2 = pd.melt(population[['FIPStxt'] + cols], 
                    id_vars=['FIPStxt'], 
                    value_vars=cols, 
                    value_name=key,
                    var_name='year')
    temp2['year'] = [x.split("_")[3] for x in temp2['year']]
    temp2['year'] = ['2020' if x == '2019' else x for x in temp2['year']]
    
    for year in temp2['year'].unique():
        df_temp = temp2[temp2['year'] == year]
        df_temp.rename(columns={'FIPStxt': 'county_fips'}, inplace=True)
        df_temp.drop(['year'], inplace=True, axis=1)
        df_temp.to_csv(os.path.join(OUTPUT_PATH, f'{key}_{year}.csv'), index=False)
        

natural_inc_year = pd.melt(population[['FIPStxt'] + rate_nat_inc_cols], 
                           id_vars=['FIPStxt'], 
                           value_vars=rate_nat_inc_cols, 
                           value_name='natural_increase_in_pop',
                           var_name='year')

natural_inc_year['year'] = [x.split("_")[3] for x in natural_inc_year['year']]
natural_inc_year['year'] = ['2020' if x == '2019' else x for x in natural_inc_year['year']]

#### Facebook

facebook = pd.read_csv(os.path.join(PATH, 'Facebook_Cornell_ERS_SocialConnectednessIndex.csv'))
facebook.rename(columns={'FIPS': 'county_fips'}, inplace=True)
facebook = facebook[['county_fips', 'sh050m', 'sh100m', 'sh500m']]

facebook.to_csv(os.path.join(OUTPUT_PATH, f'facebook_2014.csv'), index=False)

#### Social Capital

social_capital = pd.read_csv(os.path.join(PATH, 'PennState_AgriScienceDept_Religion_and_Social_Capital.csv'))
social_capital = social_capital.iloc[:, :18]
social_capital.drop(['County_Name', 'pop2014'], inplace=True, axis=1)
social_capital.rename(columns={'FIPS': 'county_fips'}, inplace=True)

social_capital.to_csv(os.path.join(OUTPUT_PATH, f'social_capital.csv'), index=False)

#### Infra

infra = pd.read_csv(os.path.join(PATH, 'HarvardData_streets_intersections.csv'))
infra.rename(columns={'geoid': 'county_fips'}, inplace=True)
infra = infra[['county_fips', 'streets_per_node_avg']]

infra.head()

infra.to_csv(os.path.join(OUTPUT_PATH, f'infra.csv'), index=False)

#### Broadband

broadband = pd.read_csv(os.path.join(PATH, 'FCC_Health_and_Broadband.csv'))
broadband.rename(columns={'geography_id': 'county_fips'}, inplace=True)
broadband = broadband[['county_fips', 'pctpopwbbacc']]

broadband.head()

broadband.to_csv(os.path.join(OUTPUT_PATH, f'broadband.csv'), index=False)
