==> 02-predicting-election-winner-by-county-first-attempt.ipynb <==
import pandas as pd
import numpy as np
import os


from sklearn.model_selection import train_test_split

from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.metrics import accuracy_score
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC
from xgboost import XGBClassifier

OUTPUT_PATH = 'outputs'

df = pd.read_csv(os.path.join(OUTPUT_PATH, 'results_2020.csv'))
df = df[['county_fips', 'winner']]

df2 = pd.read_csv(os.path.join(OUTPUT_PATH, 'rate_natural_increase_population_2020.csv'))
df3 = pd.read_csv(os.path.join(OUTPUT_PATH, 'rate_international_migration_2020.csv'))
df4 = pd.read_csv(os.path.join(OUTPUT_PATH, 'rate_domestic_migration_2020.csv'))
df5 = pd.read_csv(os.path.join(OUTPUT_PATH, 'facebook_2014.csv'))
df6 = pd.read_csv(os.path.join(OUTPUT_PATH, 'social_capital.csv')) 

merged = df.merge(df2, on='county_fips', how='left')\
           .merge(df3, on='county_fips', how='left')\
           .merge(df4, on='county_fips', how='left')\
           .merge(df5, on='county_fips', how='left')\
           .merge(df6, on='county_fips', how='left')

merged.fillna(0, inplace=True)

X = merged.drop(['county_fips', 'winner'], axis=1)
y = [1 if x == 'republican' else 0 for x in merged['winner']]

from sklearn.preprocessing import StandardScaler, MinMaxScaler

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.30, random_state=27)

scaler = MinMaxScaler()
scaler.fit(X_train)
X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)

KNN_model = KNeighborsClassifier(n_neighbors=5)
SVC_model = SVC()

SVC_model.fit(X_train, y_train)
KNN_model.fit(X_train, y_train)

SVC_prediction = SVC_model.predict(X_test)
KNN_prediction = KNN_model.predict(X_test)

print(accuracy_score(SVC_prediction, y_test))
print(accuracy_score(KNN_prediction, y_test))

xgb = XGBClassifier(n_estimators=100, objective='binary:logistic', use_label_encoder=False)
xgb.fit(X_train, y_train)
xgb_pred = xgb.predict(X_test)
print(accuracy_score(xgb_pred, y_test))
