==> 03-final-prediction-notebook.ipynb <==
## Predicting Results of US Elections on a County Level

#### Table of Contents <a name='top'></a>

- [Load Modules and Set Notebook Properties](#modules)
- [Define Path and Load Data](#load)
- [Inspect Data](#inspect)
- [Define a State Selection Function ](#stateselect)
- [Define Scaling Function](#scale)
- [Create Different Classifier Models](#create)
- [Find the Best Model](#evaluate)
- [Evaluate and Choose Models](#evaluate)
- [Predict](#predict)
- [Prepare Submission](#submit)

[go to end](#end)

#### Load Modules and Set Notebook Properties <a name='modules'></a>

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
import sys
import os

from config import ALL_STATES

from sklearn.metrics import roc_auc_score, accuracy_score, classification_report
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, StandardScaler, MinMaxScaler, Normalizer

from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC
from xgboost import XGBClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import BernoulliNB
from sklearn.tree import DecisionTreeClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.model_selection import GridSearchCV

warnings.filterwarnings('ignore')
pd.options.display.max_columns = None
sns.set_style("darkgrid")

#### Define Path and Load Data  <a name='load'></a> 

OUTPUT_PATH = 'outputs'

df = pd.read_csv(os.path.join(OUTPUT_PATH, 'results_2020.csv'))
df_ = df[['county_fips', 'state', 'winner']]

df2 = pd.read_csv(os.path.join(OUTPUT_PATH, 'rate_natural_increase_population_2020.csv'))
df3 = pd.read_csv(os.path.join(OUTPUT_PATH, 'rate_international_migration_2020.csv'))
df4 = pd.read_csv(os.path.join(OUTPUT_PATH, 'rate_domestic_migration_2020.csv'))
df5 = pd.read_csv(os.path.join(OUTPUT_PATH, 'facebook_2014.csv'))
df6 = pd.read_csv(os.path.join(OUTPUT_PATH, 'social_capital.csv')) 
df7 = pd.read_csv(os.path.join(OUTPUT_PATH, 'infra.csv')) 
df8 = pd.read_csv(os.path.join(OUTPUT_PATH, 'broadband.csv')) 

merged = df_.merge(df2, on='county_fips', how='left')\
            .merge(df3, on='county_fips', how='left')\
            .merge(df4, on='county_fips', how='left')\
            .merge(df5, on='county_fips', how='left')\
            .merge(df6, on='county_fips', how='left')\
            .merge(df7, on='county_fips', how='left')\
            .merge(df8, on='county_fips', how='left')

merged.fillna(0, inplace=True)

#### Inspect Data  <a name='inspect'></a> 

merged.head()

merged.mean()

#### Define a State Selection Function <a name='stateselect'></a> 

def select_states(states=ALL_STATES):
    
    return merged[merged['state'].isin(states)]

#### Define Scaling Function <a name='scale'></a> 

def scale_values(X_train, X_test, scaler='standard'):
    
    scaler_dict = {'standard': StandardScaler(), 
                    'minmax': MinMaxScaler(), 
                    'normal': Normalizer()}
    if scaler is None:
        return X_train, X_test
    elif scaler not in scaler_dict.keys():
        raise ValueError("Enter a valid value for scaler! Choose between 'standard', 'minmax', 'normal' or None.")
    else:
        scl = scaler_dict[scaler]
        X_train = scl.fit_transform(X_train)
        X_test = scl.transform(X_test) 
        return X_train, X_test

#### Define Classifiers

knn = KNeighborsClassifier(n_neighbors=5)
svc = SVC(probability=True)
log_reg =  LogisticRegression(penalty="l2", C=1e42, solver='liblinear')
nb = BernoulliNB(binarize=True)
dt = DecisionTreeClassifier(criterion = 'entropy', 
                            max_depth=3, 
                            random_state = 0)
xgb = XGBClassifier(n_estimators=100, 
                    objective='binary:logistic', 
                    use_label_encoder=False, 
                    eval_metric='error')

classifier_dict = {'k-Nearest Neighbors': knn,
                   'Support Vector': svc,
                   'Logistic Regression': log_reg,
                   'Naive Bayes': nb,
                   'Decision Tree': dt,
                   'XGBoost': xgb}

#### Test Classifiers without Hyperparameter Tuning

final = select_states(['ARIZONA', 'GEORGIA', 'PENNSYLVANIA', 'FLORIDA', 'OHIO', 'MICHIGAN', 'WISCONSIN'])
# final = select_states()
X = final.drop(['county_fips', 'state', 'winner', 'sh100m','sh500m',
                'Business2014','Political2014','Professional2014',
                'Labor2014','Bowling2014','Recreational2014','Golf2014',
                'Sports2014','assn2014','pvote2012','respn2010','nccs2014','sk2014'], axis=1)
y = [1 if x == 'republican' else 0 for x in final['winner']]
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.30, random_state=27)
X_train, X_test = scale_values(X_train, X_test, scaler='minmax')

print('Accuracy Scores')
print('===============')
for i in classifier_dict:
    clf = classifier_dict[i]
    clf.fit(X_train, y_train)
    pred = clf.predict(X_test)
    accuracy = accuracy_score(pred, y_test)
    print('{}: Accuracy: {:.4f}, AUC {:.4f}'.format(i, accuracy, roc_auc_score(y_test, clf.predict_proba(X_test)[:,1])))

#### Tune Classifiers

knn2 = KNeighborsClassifier()
svc2 = SVC()
log_reg2 =  LogisticRegression()
dt2 = DecisionTreeClassifier()
xgb2 = XGBClassifier()

knn_params = {'n_neighbors': list(range(1,20))}
svc_params = {'C': np.logspace(-3, 2, 6), 
              'gamma': np.logspace(-3, 2, 6),
              'probability': [True]}
logreg_params = {'solver': ['newton-cg', 'lbfgs', 'liblinear'],
                 'penalty': ['l2'],
                 'C': [100, 10, 1.0, 0.1, 0.01]}
dt_params = {'criterion': ['gini', 'entropy'],
             'max_depth': range(1,10),
             'min_samples_split': range(1,10),
             'min_samples_leaf': range(1,10)}
xgb_params = {"learning_rate": [0.05, 0.10],
              'n_estimators': [10, 20, 40, 100],
              'objective': ['binary:logistic'],
              'use_label_encoder': [False],
              'eval_metric': ['auc']}

classifier_dict2 = {'k-Nearest Neighbors': [knn2, knn_params],
                    'Support Vector': [svc2, svc_params],
                    'Logistic Regression': [log_reg2, logreg_params],
                    'Decision Tree': [dt2, dt_params],
                    'XGBoost': [xgb2, xgb_params]}

print('Hyperparameter Optimization')
print('===========================')
for i in classifier_dict2:
    print(i)
    clf = classifier_dict2[i][0]
    params = classifier_dict2[i][1]
    grid = GridSearchCV(clf, params, cv=3, scoring='accuracy')
    grid.fit(X_train, y_train)
    print(f'Best Parameters: {grid.best_params_}')
    print('Accuracy on Training: {:.4f}'.format(grid.best_score_))
    roc_auc_score
    
    best_model = grid.best_estimator_
    pred = best_model.predict(X_test)
    accuracy = accuracy_score(pred, y_test)
    print('Accuracy on Test: {:.4f}'.format(accuracy))
    print('AUC on Test: {:.4f}'.format(roc_auc_score(y_test, best_model.predict_proba(X_test)[:,1])))
    print('+++++++++++++++++++++++++++')

X.columns
